requirements.txt
aiohttp>=3.8.0
pydantic>=2.0.0
bleak>=0.20.0
paho-mqtt>=2.0.0